--- src/components/Modal/NotificationModal.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const NotificationModal = () => {
-  return <div>알림창</div>;
-};
-
-export default NotificationModal;
