+++ src/components/Modal/_CommonStyle.scss
@@ -4,8 +4,8 @@
   position: absolute;
   width: 465px;
   height: fit-content;
-  max-height: 650px;
-  right: 102px;
+  max-height: 680px;
+  right: 100px;
   top: 20px;
   z-index: 10;
   border-radius: 10px;
